[#217] update env vars for dev Kinesis
IAM permissions have been granted for new dev Kinesis streams. This
updates env vars so everything works out of the box.

It was not necessary to set PSC_STREAM_API_KEY. This is used when
subscribing to a PSC stream, not when publishing to a PSC stream, as far
as I know. That code has not been run in some time (upcoming work).

It was necessary to set BODS_S3_BUCKET_NAME. However, similar to
Ingester DK, this seemingly requires read-write permissions, when
running the whole ingest-bulk process. Running only the ingest-snapshots
part, however, still requires it to be set, but read-only is sufficient.
So, I'm setting it, but leaving it read-only for dev at present.