[#243] extend StreamIngester to crash on invalid timepoint
If the stream timepoint is invalid, such as if it is too far in the
past, it cannot be used. In such a case, the stored timepoint is cleared
from Redis, and the program crashes.

This approach is chosen rather than clearing the timepoint and
continuing, since this should usually be an exceptional failure case,
which merits bubbling up through whichever monitoring method is being
used. Restarting the program is left to whatever is monitoring it.